requirements.txt
ics==0.7.2
requests==2.31.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0